fix ruff
